update device log migration

+++ database/migrations/2022_07_28_164544_create_device_logs_table.php
@@ -16,7 +16,7 @@ class CreateDeviceLogsTable extends Migration
         Schema::create('device_logs', function (Blueprint $table) {
             $table->id();
             $table->foreignId('device_id');
-            $table->float('status')->default(0);
+            $table->integer('status')->default(0);
             $table->enum('is_auto', ['Yes', 'No'])->default('No');
             $table->double('temperature', 5, 2)->default(0);
             $table->timestamp('timestamp');
